backend requirements

+++ backend/src/routes/v1/shifts/shiftController.ts
@@ -3,10 +3,13 @@ import * as shiftUsecase from "../../../usecases/shiftUsecase";
 import { errorHandler } from "../../../shared/functions/error";
 import {
   ICreateShift,
+  IErrorResponse,
+  IPublishShift,
   ISuccessResponse,
   IUpdateShift,
 } from "../../../shared/interfaces";
 import moduleLogger from "../../../shared/functions/logger";
+import Shift from "../../../database/default/entity/shift";
 
 const logger = moduleLogger("shiftController");
 
@@ -22,7 +25,7 @@ export const find = async (req: Request, h: ResponseToolkit) => {
     };
     return res;
   } catch (error) {
-    logger.error(error.message)
+    logger.error(error.message);
     return errorHandler(h, error);
   }
 };
@@ -39,7 +42,7 @@ export const findById = async (req: Request, h: ResponseToolkit) => {
     };
     return res;
   } catch (error) {
-    logger.error(error.message)
+    logger.error(error.message);
     return errorHandler(h, error);
   }
 };
@@ -48,6 +51,37 @@ export const create = async (req: Request, h: ResponseToolkit) => {
   logger.info("Create shift");
   try {
     const body = req.payload as ICreateShift;
+
+    if (body.startTime && body.endTime && body.date) {
+      // check if shift overlap others
+      const overlappingShifts = await shiftUsecase.findOverlap(
+        body.startTime,
+        body.endTime,
+        body.date
+      );
+      if (overlappingShifts.length > 0) {
+        return {
+          error: "Cannot create shift",
+          statusCode: 400,
+          message: "Shift overlapped with other Shift",
+        } as IErrorResponse;
+      }
+
+      // check if the date is already been published
+      const date = new Date(body.date);
+      const publishedShifts = await shiftUsecase.findPublishedShiftsInTheWeek(
+        date
+      );
+
+      if (publishedShifts.length > 0) {
+        return {
+          error: "Cannot create shift",
+          statusCode: 400,
+          message: "The date is on the week that has already published",
+        } as IErrorResponse;
+      }
+    }
+
     const data = await shiftUsecase.create(body);
     const res: ISuccessResponse = {
       statusCode: 200,
@@ -56,7 +90,7 @@ export const create = async (req: Request, h: ResponseToolkit) => {
     };
     return res;
   } catch (error) {
-    logger.error(error.message)
+    logger.error(error.message);
     return errorHandler(h, error);
   }
 };
@@ -67,6 +101,32 @@ export const updateById = async (req: Request, h: ResponseToolkit) => {
     const id = req.params.id;
     const body = req.payload as IUpdateShift;
 
+    // check if shift overlap others
+    if (body.startTime && body.endTime && body.date) {
+      const overlappingShifts = await shiftUsecase.findOverlap(
+        body.startTime,
+        body.endTime,
+        body.date
+      );
+      if (overlappingShifts.length > 0) {
+        return {
+          error: "Cannot update shift",
+          statusCode: 400,
+          message: "Shift overlapped with other Shift",
+        } as IErrorResponse;
+      }
+    }
+
+    // check if shift has been published
+    const shift = await shiftUsecase.findById(id);
+    if (shift.isPublished) {
+      return {
+        error: "Cannot update shift",
+        statusCode: 400,
+        message: "Shift has already been published",
+      } as IErrorResponse;
+    }
+
     const data = await shiftUsecase.updateById(id, body);
     const res: ISuccessResponse = {
       statusCode: 200,
@@ -75,7 +135,7 @@ export const updateById = async (req: Request, h: ResponseToolkit) => {
     };
     return res;
   } catch (error) {
-    logger.error(error.message)
+    logger.error(error.message);
     return errorHandler(h, error);
   }
 };
@@ -84,6 +144,17 @@ export const deleteById = async (req: Request, h: ResponseToolkit) => {
   logger.info("Delete shift by id");
   try {
     const id = req.params.id;
+
+    // check if shift has been published
+    const shift = await shiftUsecase.findById(id);
+    if (shift.isPublished) {
+      return {
+        error: "Cannot delete shift",
+        statusCode: 400,
+        message: "Shift has already been published",
+      } as IErrorResponse;
+    }
+
     const data = await shiftUsecase.deleteById(id);
     const res: ISuccessResponse = {
       statusCode: 200,
@@ -92,7 +163,45 @@ export const deleteById = async (req: Request, h: ResponseToolkit) => {
     };
     return res;
   } catch (error) {
-    logger.error(error.message)
+    logger.error(error.message);
+    return errorHandler(h, error);
+  }
+};
+
+export const publish = async (req: Request, h: ResponseToolkit) => {
+  logger.info("Publish shift(s)");
+  try {
+    const body = req.payload as IPublishShift;
+
+    // check the which week is going to be published
+    const startDate = body.startDate;
+    const endDate = body.endDate;
+
+    // get all shifts whithin the week
+    const shifts = await shiftUsecase.findShiftsBetweenDate(startDate, endDate);
+
+    if (shifts.length === 0) {
+      return {
+        error: "Cannot publish shift",
+        statusCode: 400,
+        message: "No shifts found between the date",
+      } as IErrorResponse;
+    }
+
+    // update isPublished field in every shift
+    shifts.forEach((shift: Shift) => {
+      shift.isPublished = true;
+    });
+
+    const data = await shiftUsecase.save(shifts);
+    const res: ISuccessResponse = {
+      statusCode: 200,
+      message: "Publish shift successful",
+      results: data,
+    };
+    return res;
+  } catch (error) {
+    logger.error(error.message);
     return errorHandler(h, error);
   }
 };
